FoodAI/nutrition-video-analysis/terraform/docker/requirements.txt
# AWS SDK
boto3>=1.34.0

# API & Config
pydantic-settings==2.0.3

# Deep Learning & Computer Vision
torch==2.4.1
torchvision==0.19.1
transformers>=4.41.0,<5.0.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
einops==0.7.0
timm==0.9.12
huggingface-hub>=0.34.0
httpx>=0.28.1
accelerate>=1.10.0

# SAM2 Requirements
hydra-core==1.3.2
iopath==0.1.10
tqdm==4.66.1

# Depth / Visualization Requirements
matplotlib==3.7.5
h5py==3.10.0

# RAG & NLP
sentence-transformers==3.3.1
scikit-learn==1.5.2
faiss-cpu==1.7.4

# External APIs
google-generativeai==0.3.1
google-genai>=0.8.0
